requirements.txt
build==0.10.0
twine==4.0.2
PyYAML==6.0
GitPython==3.1.31
xh-dual-layer-app-engine==0.0.2